BAU: Refactor auth types into concern and add `current_user`
